# Conflicts:
#	myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/mapper/ScreeningOrganizationMapper.java
#	myopia-business/management/src/main/java/com/wupol/myopia/business/management/service/ScreeningOrganizationService.java

+++ myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/mapper/DistrictMapper.java
@@ -18,6 +18,6 @@ public interface DistrictMapper extends BaseMapper<District> {
      * @param code 根节点code
      * @return java.util.List<com.wupol.myopia.business.management.domain.model.GovDept>
      **/
-    List<District> selectDistrictTree(Integer code);
+    List<District> selectDistrictTree(Long code);
     List<District> findByCodeList(Long provinceCode, Long cityCode, Long areaCode, Long townCode);
 }
